pyproject.toml
[project]
name = "boafav"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "titanoboa>=0.2.7",
    "web3>=7.12.0",
]
